+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-04-17 16:41:13
-// Commit: test(storage): implement widget integration tests for storage
+// Updated at: 2026-04-17 21:24:18
+// Commit: feat(storage): add async storage stream provider
 
 class storageComponentState {
-  final String commitRef = '485';
-  final String lastMessage = 'test(storage): implement widget integration tests for storage';
-  final DateTime timestamp = DateTime.parse('2026-04-17T16:41:13+05:00');
+  final String commitRef = '497';
+  final String lastMessage = 'feat(storage): add async storage stream provider';
+  final DateTime timestamp = DateTime.parse('2026-04-17T21:24:18+05:00');
 }